+++ src/RestApi/Device/Screenshot/IDeviceScreenshotFilter.ts
@@ -1,0 +1,7 @@
+interface IDeviceScreenshotFilter {
+	takenSince?: Date;
+	takenUntil?: Date;
+	limit?: number;
+	descending?: boolean;
+}
+export default IDeviceScreenshotFilter;